% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_dir_role.R
\docType{class}
\name{az_directory_role}
\alias{az_directory_role}
\title{Directory role}
\format{
An R6 object of class \code{az_directory_role}, inheriting from \code{az_object}.
}
\description{
Class representing a role in Azure Active Directory.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for this role.
\item \code{type}: always "directory role" for a directory role object.
\item \code{properties}: The item properties.
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new object. Do not call this directly; see 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete this item. By default, ask for confirmation first.
\item \code{update(...)}: Update the item's properties in Microsoft Graph.
\item \code{do_operation(...)}: Carry out an arbitrary operation on the item.
\item \code{sync_fields()}: Synchronise the R object with the item metadata in Microsoft Graph.
}
}

\section{Initialization}{

Currently support for directory roles is limited. Objects of this class should not be initialized directly.
}

\seealso{
\link{ms_graph}, \link{az_user}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-1.0}{REST API reference}
}
