% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_graph.R
\docType{class}
\name{az_graph}
\alias{az_graph}
\title{Azure Active Directory Graph}
\format{An R6 object of class \code{az_graph}.}
\usage{
az_graph
}
\description{
Base class for interacting with Azure AD Graph API.
}
\section{Methods}{

\itemize{
\item \code{new(tenant, app, ...)}: Initialize a new ARM connection with the given credentials. See 'Authentication` for more details.
\item \code{create_app(name, ..., password=NULL, create_service_principal=TRUE)}: Creates a new registered app in Azure Active Directory. By default the app will have a randomly generated strong password with a duration of 1 year, and an associated service principal will also be created.
\item \code{get_app(app_id, object_id)}: Retrieves an existing registered app, via either its app ID or object ID.
\item \code{delete_app(app_id, object_id, confirm=TRUE)}: Deletes an existing registered app. Any associated service principal will also be deleted.
\item \code{create_service_principal(app_id, ...)}: Creates a service principal for a registered app.
\item \code{get_service_principal()}: Retrieves an existing service principal.
\item \code{delete_service_principal()}: Deletes an existing service principal.
}
}

\section{Authentication}{

The recommended way to authenticate with Azure AD Graph is via the \link{create_graph_login} function, which creates a new instance of this class.

To authenticate with the \code{az_graph} class directly, provide the following arguments to the \code{new} method:
\itemize{
\item \code{tenant}: Your tenant ID. This can be a name ("myaadtenant"), a fully qualified domain name ("myaadtenant.onmicrosoft.com" or "mycompanyname.com"), or a GUID.
\item \code{app}: The client/app ID to use to authenticate with Azure Active Directory. The default is to login interactively using the Azure CLI cross-platform app, but it's recommended to supply your own app credentials if possible.
\item \code{password}: if \code{auth_type == "client_credentials"}, the app secret; if \code{auth_type == "resource_owner"}, your account password.
\item \code{username}: if \code{auth_type == "resource_owner"}, your username.
\item \code{auth_type}: The OAuth authentication method to use, one of "client_credentials", "authorization_code", "device_code" or "resource_owner". See \link{get_azure_token} for how the default method is chosen, along with some caveats.
\item \code{host}: your Azure AD Graph host. Defaults to \code{https://graph.windows.net/}. Change this if you are using a government or private cloud.
\item \code{aad_host}: Azure Active Directory host for authentication. Defaults to \code{https://login.microsoftonline.com/}. Change this if you are using a government or private cloud.
\item \code{config_file}: Optionally, a JSON file containing any of the arguments listed above. Arguments supplied in this file take priority over those supplied on the command line. You can also use the output from the Azure CLI \code{az ad sp create-for-rbac} command.
\item \code{token}: Optionally, an OAuth 2.0 token, of class \link{AzureToken}. This allows you to reuse the authentication details for an existing session. If supplied, all other arguments will be ignored.
}
}

\examples{
\dontrun{

# start a new Resource Manager session
gr <- az_graph$new(tenant="myaadtenant.onmicrosoft.com", app="app_id", password="password")

# authenticate with credentials in a file
gr <- az_graph$new(config_file="creds.json")

# authenticate with device code
gr <- az_graph$new(tenant="myaadtenant.onmicrosoft.com", app="app_id", auth_type="device_code")

# retrieve a registered app
gr$get_app(app_id="myappid")

# create a new app and associated service principal, set password duration to 10 years
app <- gr$create_app("mynewapp", password_duration=10)

svc <- gr$get_service_principal(app_id=app$properties$appId)

# delete the app
gr$delete_app(app_id=app$properties$appId)
# ... but better to call the object's delete method directly
app$delete()

}
}
\seealso{
\link{create_graph_login}, \link{get_graph_login}

\href{https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-graph-api}{Azure AD Graph overview},
\href{https://docs.microsoft.com/en-au/previous-versions/azure/ad/graph/api/api-catalog}{REST API reference}
}
\keyword{datasets}
