% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_svc_principal.R
\docType{class}
\name{az_service_principal}
\alias{az_service_principal}
\title{Service principal in Azure Active Directory}
\format{An R6 object of class \code{az_service_principal}.}
\usage{
az_service_principal
}
\description{
Base class representing an AAD service principal.
}
\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new service principal object. See 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete a service principal. By default, ask for confirmation first.
}
}

\section{Initialization}{

The recommended way to create a new service principal object is via the \code{create_service_principal} and \code{get_service_principal} methods of the \link{az_graph} and \link{az_app} classes, which will handle the details automatically.
}

\seealso{
\link{az_graph}, \link{az_app}
}
\keyword{datasets}
