% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_public_fields}
\alias{format_public_fields}
\alias{format_public_methods}
\title{Format a Microsoft Graph or Azure object}
\usage{
format_public_fields(env, exclude = character(0))

format_public_methods(env)
}
\arguments{
\item{env}{An R6 object's environment for printing.}

\item{exclude}{Objects in \code{env} to exclude from the printout.}
}
\description{
Miscellaneous functions for printing Microsoft Graph and Azure R6 objects
}
\details{
These are utilities to aid in printing R6 objects created by this package or its descendants. They are not meant to be called by the user.
}
