% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_list}
\alias{named_list}
\alias{is_empty}
\title{Miscellaneous utility functions}
\usage{
named_list(lst = NULL, name_fields = "name")

is_empty(x)
}
\arguments{
\item{lst}{A named list of objects.}

\item{name_fields}{The components of the objects in \code{lst}, to be used as names.}

\item{x}{For \code{is_empty}, An R object.}
}
\value{
For \code{named_list}, the list that was passed in but with names. An empty input results in a \emph{named list} output: a list of length 0, with a \code{names} attribute.

For \code{is_empty}, whether the length of the object is zero (this includes the special case of \code{NULL}).
}
\description{
Miscellaneous utility functions
}
\details{
\code{named_list} extracts from each object in \code{lst}, the components named by \code{name_fields}. It then constructs names for \code{lst} from these components, separated by a \code{"/"}.
}
