% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_app.R
\docType{class}
\name{az_app}
\alias{az_app}
\title{Registered app in Azure Active Directory}
\format{An R6 object of class \code{az_app}.}
\usage{
az_app
}
\description{
Base class representing an AAD app.
}
\section{Methods}{

\itemize{
\item \code{new(token, tenant, ...)}: Initialize a new app object. See 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete an app. By default, ask for confirmation first.
\item \code{update(...)}: Update the details for an app.
\item \code{sync_fields()}: Synchronise the R object with the app data in Azure Active Directory.
\item \code{create_service_principal(...)}: Create a service principal for this app, by default in the current tenant.
\item \code{get_service_principal()}: Get the service principal for this app.
\item \code{delete_service_principal(confirm=TRUE)}: Delete the service principal for this app. By default, ask for confirmation first.
}
}

\keyword{datasets}
