% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_group.R
\docType{class}
\name{az_group}
\alias{az_group}
\title{Group in Azure Active Directory}
\format{An R6 object of class \code{az_group}.}
\usage{
az_group
}
\description{
Base class representing an AAD group.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for this group.
\item \code{properties}: The group properties.
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new group object. Do not call this directly; see 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete a group. By default, ask for confirmation first.
\item \code{update(...)}: Update the group information in Azure Active Directory.
\item \code{sync_fields()}: Synchronise the R object with the app data in Azure Active Directory.
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{create_group} and \code{get_group} methods of the \link{ms_graph} and \link{az_app} classes. Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to create the actual group.
}

\seealso{
\link{ms_graph}, \link{az_app}, \link{az_user}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-beta}{REST API reference}
}
\keyword{datasets}
