% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_user.R
\docType{class}
\name{az_user}
\alias{az_user}
\title{User in Azure Active Directory}
\format{An R6 object of class \code{az_user}.}
\usage{
az_user
}
\description{
Base class representing an AAD user account.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for this user.
\item \code{properties}: The user properties.
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new user object. Do not call this directly; see 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete a user account. By default, ask for confirmation first.
\item \code{update(...)}: Update the user information in Azure Active Directory.
\item \code{sync_fields()}: Synchronise the R object with the app data in Azure Active Directory.
\item `reset_password(password=NULL, force_password_change=TRUE): Resets a user password. By default the new password will be randomly generated, and must be changed at next login.
\item \code{list_group_memberships(direct_only=TRUE, id_only=TRUE)}: List the groups this user is a member of. Set \code{direct_only=FALSE} to get a \emph{transitive} list of memberships, ie including groups that the user's groups are members of. Set \code{id_only=TRUE} to return only a vector of group IDs (the default), or \code{id_only=FALSE} to return a list of group objects (which will be slow for a transitive list).
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{create_user} and \code{get_user} methods of the \link{ms_graph} and \link{az_app} classes. Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to create the actual user account.
}

\seealso{
\link{ms_graph}, \link{az_app}, \link{az_group}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-beta}{REST API reference}
}
\keyword{datasets}
