% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{call_batch_endpoint}
\alias{call_batch_endpoint}
\title{Call the Graph API batch endpoint}
\usage{
call_batch_endpoint(token, requests = list(), depends_on = list(),
  api_version = getOption("azure_graph_api_version"))
}
\arguments{
\item{token}{An Azure OAuth token, of class \link{AzureToken}.}

\item{requests}{A list of \link{graph_request} objects, representing individual requests to the Graph API.}

\item{depends_on}{An optional named list indicating which requests have dependencies. See the examples below.}

\item{api_version}{The API version to use, which will form part of the URL sent to the host.}
}
\value{
A list containing the responses to each request. Each item has components \code{id} and \code{status} at a minimum. It may also contain \code{headers} and \code{body}, depending on the specifics of the request.
}
\description{
Call the Graph API batch endpoint
}
\details{
Use this function to combine multiple requests into a single HTTPS call. This can save significant network latency.
}
\examples{
\dontrun{

req1 <- graph_request$new("me")

# a new email message in Outlook
req2 <- graph_request$new("me/messages",
    body=list(
        body=list(
            content="Hello from R",
            content_type="text"
        ),
        subject="Hello",
        toRecipients="bob@example.com"
    ),
    http_verb="POST"
)

# messages in drafts
req3 <- graph_request$new("me/mailFolders/drafts/messages")

# depends_on signifies that request 3 depends on request 2
call_batch_endpoint(token, list(req1, req2, req3), depends_on=list("3"=2))

}
}
\seealso{
\link{graph_request}, \link{call_graph_endpoint}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/json-batching}{Batch endpoint documentation}

\href{https://docs.oasis-open.org/odata/odata-json-format/v4.01/odata-json-format-v4.01.html#sec_BatchRequestsandResponses}{OData documentation on batch requests}
}
