% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_class_directory.R
\name{find_class_generator}
\alias{find_class_generator}
\title{Find the R6 class for a Graph object}
\usage{
find_class_generator(props, type_filter = NULL, default_generator = ms_object)
}
\arguments{
\item{props}{A list of object properties, generally the result of a Graph API call.}

\item{type_filter}{An optional vector of types by which to filter the result.}

\item{default_generator}{The default class generator to use, if a match couldn't be found.}
}
\value{
An R6 class generator for the appropriate AzureGraph class. If no matching R6 class could be found, the default generator is returned. If \code{type_filter} is provided, but the matching R6 class isn't in the filter, NULL is returned.
}
\description{
Find the R6 class for a Graph object
}
\details{
This function maps Graph objects to AzureGraph classes.
}
