% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_app}
\alias{is_app}
\alias{is_service_principal}
\alias{is_user}
\alias{is_group}
\alias{is_directory_object}
\title{Informational functions}
\usage{
is_app(object)

is_service_principal(object)

is_user(object)

is_group(object)

is_directory_object(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
A boolean.
}
\description{
These functions return whether the object is of the corresponding class.
}
